Se ha creado una clase AbstractPersistable propia ya que el método getId era protected y Thymeleaf fallaba. Se ha implementado una cache de nivel 1 a las clases de Services para certification y familyprofessional, faltaría configurar bien el archivo con los tiempo de vida, etc. Se han eliminado los repositórios y se ha creado una interfaz que extiende a CrudRepository para cada Entity. Spring se encarga de realizar una implementación genérica para cada uno.

+++ src/main/java/com/acme/services/UserService.java
@@ -7,5 +7,5 @@ import com.acme.model.user.User;
 
 public interface UserService {
 	public List<User> getAllCompanyWithId();
-	public User getUserById(Integer id);
+	public User getUserById(Long id);
 }
